Change size data type

+++ lib_artiexh-data-jpa/src/main/java/com/artiexh/data/jpa/entity/embededmodel/Size.java
@@ -10,6 +10,6 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Size {
-	private Integer width;
-	private Integer height;
+	private Float width;
+	private Float height;
 }
